WIP PostDeleteButton component
