+++ app/components/PostsList.tsx
@@ -1,5 +1,5 @@
 import prisma from '@/app/lib/db';
-import { revalidatePath } from 'next/cache';
+import PostDeleteButton from '@/app/components/PostDeleteButton'
 
 type PostsListProps = {
   userId?: number,
@@ -13,14 +13,6 @@ export default async function PostsList({ userId, page = 0 } : PostsListProps) {
 
   // await new Promise((resolve) => setTimeout(resolve, 1000))
 
-  const deletePost = async (formData: FormData) => {
-    "use server";
-
-    await prisma.post.deleteMany({ where : { id: Number(formData.get("postId") as string) }})
-
-    revalidatePath("/posts")
-  }
-
   return (
     <>
       <ul>
@@ -29,10 +21,7 @@ export default async function PostsList({ userId, page = 0 } : PostsListProps) {
             <h2 className='font-bold'>{post.title}</h2>
             <p>{post.body}</p>
             <small>userId: {post.userId}</small> <small>postId: {post.id}</small>
-            <form action={deletePost}>
-              <input type="hidden" name="postId" value={post.id} />
-              <button className='bg-red-300'>Delete</button>
-            </form>
+            <PostDeleteButton postId={post.id} />
           </li>
         ))}
       </ul>
